daemon: change default policy enforcement behavior for non-K8s envs
Previously, in an environment where Cilium was not running in tandem with
Kubernetes, policy enforcement would be enabled for all endpoints if a  rule
was added to the policy repository even if said rule did not apply to an
endpoint. This changes that behavior for Cilium running without Kubernetes to
enable or disable policy enforcment as is done when Cilium is ran with
Kubernetes; policy enforcement is only enabled for an endpoint when a rule
selects that endpoint when the daemon is ran in 'default' policy enforcement
mode.

Signed-off by: Ian Vernon <ian@cilium.io>